rosbag2 confirmed working; added 'strip' build option
added 'registerPluginlibPlugin' for registration of pluginlib plugins
such as rosbag2_storage_mcap.

rosbag2 recorder confirmed working in a downstream repo.

TODO: rosbag2_storage_mcap is a pluginlib plugin, which means it
requires a bunch of text files to be installed in order to be "found"
and used. We need a way to register these kinds of things (steps to
be installed by the child builder) without the downstream user
needing to know about and manually install them using their builder.

something like: 'zigros.installAllPackages(b);'